Issue 5808: SLTS - Add more coverage.
